Services/requirements.txt
bson
pymongo
python-dotenv